Changing byte to int
C# doesnt like implicit type casts, and it seems other stuff uses ints too it should be fine

+++ Content.Shared/_RMC14/Weapons/Ranged/GetGunAPStacksModifierEvent.cs
@@ -2,4 +2,4 @@
 namespace Content.Shared._RMC14.Weapons.Ranged;
 
 [ByRefEvent]
-public record struct GetGunAPStacksModifierEvent(byte Stacks);
+public record struct GetGunAPStacksModifierEvent(int Stacks);
